+++ app/src/main/java/com/example/travelplus/App.java
@@ -1,0 +1,4 @@
+package com.example.travelplus;
+
+public class App {
+}
