--- app/src/main/java/com/example/travelplus/IsFirstResponse.java
@@ -1,13 +1,0 @@
-package com.example.travelplus;
-
-public class IsFirstResponse {
-    public boolean isFirst;
-    public int resultCode;
-    public String resultMessage;
-    public int courseId;
-    public String title;
-    public String area;
-    public String startDate;
-    public String endDate;
-    public String meansTp;
-}
