+++ app/src/main/java/com/example/travelplus/login/KakaoResponse.java
@@ -1,0 +1,6 @@
+package com.example.travelplus.login;
+
+import androidx.fragment.app.Fragment;
+
+public class KakaoLoginFragment extends Fragment {
+}
